libteec: allow passing settings to session open
Add a new function, TEEC_OpenSession2, that allows a developer to pass
configuration parameters in the form of distinct settings. One of the new
settings is for attaching an arbitrary pointer to the session. This is
useful when an OCALL arrives from the TA and the handler requires
contextual information to proceed that changes depending on which session
the OCALL arrived through.

Having multiple settings in this manner helps reduce the number of
auxiliary functions necessary. Were new functionality to be added in the
future, no new functions would need to be introduced. Instead, one would
only require a new setting.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>