libteec: implement OCALL support during session open & function invocation
OCALLs allow a TA to invoke functions on their CA with parameters, if
desired.

The flow begins when a CA calls TEEC_InvokeCommand. If the TEE context was
initialized with the OCALL setting, libteec includes an additional
parameter, the OCALL parameter, in the function invocation IOCTL. The
presence of the OCALL parameter lets the kernel driver know that an OCALL
may result from the invocation.

If an OCALL does arrive from the TA, the OCALL parameter includes
information about the OCALL, including the ID of the function that libteec
must handle. These are: allocate shared memory, free shared memory, and
invoke a function on the CA.

If either of the first two functions arrives at libteec, the library
handles these on behalf of the CA, allocating and freeing shared memory as
necessary.

When the third function arrives, libteec processes the OCALL's parameters.
These will have temporarily replaced the parameters of the original
function invocation. Additionally, the 'func' element of the IOCTL
parameters will have been modified to carry the command ID that the TA
requests that the CA execute on its behalf.

The library passes this ID along with the parameters and arbitrary data
pointers configured via the settings API to the CA-provided OCALL handler.

After the handler is finished processing the request, libteec performs
minor post-processing on the parameters and calls back into the driver to
let it know that the OCALL has been handled.

It is possible for a TA to invoke multiple OCALLs in the same originating
function invocation. Additionally, the same mechanism detailed above
applies to when a CA calls TEEC_OpenSession or TEEC_OpenSession2. That is,
the TA may invoke one or more functions in sequence on its CA during
session open.

Lastly, the CA-provided OCALL handler is executed in the context of the
thread that originally called TEEC_OpenSession, TEEC_OpenSession2, or
TEEC_InvokeCommand. That is, when a TA invokes a function on its CA, these
functions call into the CA-provided OCALL handler.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>
Reviewed-by: Jerome Forissier <jerome@forissier.org>